requirements.txt
fastapi==0.115.0
python-dotenv==1.0.1
pydantic==2.9.2
loguru==0.5.3
asyncpg==0.29.0
openai==1.50.2
uvicorn==0.31.0
starlette==0.38.5
apscheduler==3.10.4